+++ src/display_drivers/ncurses_display_driver.cpp
@@ -2,6 +2,7 @@
 
 #include "../document/url_line_element.hpp"
 #include "../document/code_line_element.hpp"
+#include "../document/bold_line_element.hpp"
 
 #include <cassert>
 
@@ -223,6 +224,8 @@ int NcursesDisplayDriver::render(Document* doc, const int & line_offset) const {
 				} else {
 					wattron(_display_window, A_REVERSE);
 				}
+			} else if (dynamic_cast<BoldLineElement*>(l)) { // Bold element
+				wattron(_display_window, A_BOLD);
 			}
 			if (tmp_str.size() < _width - cursor_x) {
 				if (bounds_check(_display_window, cursor_y, cursor_x)) {
@@ -248,8 +251,9 @@ int NcursesDisplayDriver::render(Document* doc, const int & line_offset) const {
 				} else {
 					wattroff(_display_window, A_REVERSE);
 				}
+			} else if (dynamic_cast<BoldLineElement*>(l)) { // Bold element
+				wattroff(_display_window, A_BOLD);
 			}
-
 		}
 		switch (p->level()) {
 			case Paragraph::Level::Title1: