+++ src/document/bold_line_element.cpp
@@ -1,0 +1,4 @@
+#include "bold_line_element.hpp"
+
+BoldLineElement::BoldLineElement(const std::string & content) :
+	LineElement(content) {}