+++ ChangeLog.md
@@ -9,11 +9,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * src/parsers/md_parser.cpp (MdParser::parse_line): parse inline code snippets (using regex)
 * src/parsers/md_parser.cpp (MdParser::parse): detect quote paragraphs
+* src/parsers/md_parser.cpp (MdParser::parse_line): parse bold elements
 * src/documents/paragraph.cpp (class definition): add Paragraph::Level::Quote
 * src/display_drivers/ncurses_display_driver.cpp (N[...]::render): render QuoteParagaph
+* src/display_drivers/ncurses_display_driver.cpp (N[...]::render): render BoldLineElement
 * CMakeLists.txt (add_executable): add src/document/quote_paragraph.cpp
+* CMakeLists.txt (add_executable): add src/document/bold_line_element.cpp
 * src/document/quote_paragraph.hpp: new file
 * src/document/quote_paragraph.cpp: new file
+* src/document/bold_line_element.hpp: new file
+* src/document/bold_line_element.cpp: new file
 
 ### Changed